GIN index support
For GIN indexing, we degrade all mocs to fixed level MOC_GIN_ORDER (5),
and for each pixel, the index stores the set of all mocs where this
pixel appears in.

The query is degraded to the same level, and the GIN index is queried
for these pixels.

For intersects (&&) operations, a match is found if at least one pixel
matches. Since we are operating on degraded input, rechecking against
the original moc is required.