Implement restriction selectivity estimators for <@(spoint, scircle)
This implements restriction selectivity estimation for the <@ @> !<@ !@>
family of operators on spoint and scircle. The selectivity is estimated
to be (area of sphere circle) / (4 pi).

Queries like `select * from sky where sky.star <@ scircle(const, radius)`
will be able to properly estimate if using an index is appropriate
depending on the size of radius.

Secondly, a function spoint_dwithin(p1 spoint, p2 spoint, radius float8)
is added that effectively returns `p1 <-> p2 <= radius`. But other than
this two-operator expression, it has GIST index support so the optimizer
can rewrite it to either `p1 <@ scircle(p2, radius)` or `p2 <@
scircle(p1, radius)`, i.e. it is symmetric in the first two arguments.

This allows efficient matching queries without the user having to encode
the join ordering in the query.

On PostgreSQL 10/11, the spoint_dwithin function is created, but without
the GIST support since that only appeared in PG12.

The file expected/selectivity_1.out is used on PG10/11; it has <@
flipped around to @> in some plans.