Migrate doc generation to docbook/fop (postgresql approach)
The document generation was redesigned to use the same approach as postgresql
does. It uses fop for printable doc generation and xsltproc for html
generation. Postgresql stylesheets are used (with some minor changes).

Fixed version propagation when generating the doc. Now, the version is
placed in Makefile.common.mk file that is used in doc generation
process.

The following make rules were implemented:
- make pdf (generates printable doc in A4 and US Letter format)
- make pg_sphere-<VERSION>-A4.pdf (generate the printable doc in A4)
- make pg_sphere-<VERSION>-US.pdf (generate the printable doc in USLetter)
- make html (generates single-page and multi-page html)
- make html-singlepage (generate single-page html only)
- make html-multipage (generate multi-page html only)