Disable install/installcheck targets in doc/Makefile
Remove inclusion of 'pg_config --pgxs' makefile because it is not
required to create the documentation. This makefile was designed to be
included in top-level extension makefile. Inclusion in the doc/Makefile
results into some unexpected behaviour when using 'install' or 'all' rules.

Remove gmake -C doc install in .travis.yml because this target didn't
do anything. I guess that the installation of the manual should be
implemented in the top-level extension Makefile file, not in the
doc/Makefile.