Merge index fetch support for spoint3 into main files
So far, the function backing index only fetches on spoint3 gist indexes
was already always compiled, but only added to the operator class for
PG9.5+, and only when special code in the Makefile was uncommented. We
are well past 9.5, and the feature works just fine, so load it
unconditionally (assuming no one used it in production yet, ADD FUNCTION
will fail when executed twice). Version bumped since we add a new
function to the opclass.