Now more carefully nulling out null inputs in epoch_prop.
This is so we do not invent values for pm, parallax, or rv.  We
are actually a bit over-cautious and invalidate both PMs even if
only one is missing.  This is because PMs mix, and hence there are traces
of the invented 0 in the other component of the PM.  Similarly,
when the parallax is missing or bad, the RV would be heavily contaminated;
in these cases, we copy through the original RV, as it may still be useful
and certainly will not be grossly wrong.

(sorry about a few whitespace diffs introduced by killing trailing whitespace
in sql/epochprop.sql and src/epochprop.c)